[REFACTOR]: Refactor folder structure
