--- src/components/pages/Admin/Admin.jsx
@@ -1,11 +1,0 @@
-import React from "react";
-
-const Admin = () => {
-  return (
-    <div className="text-5xl">
-      <h1>Admin - Protacted Route</h1>
-    </div>
-  );
-};
-
-export default Admin;
