+++ src/components/pages/builder/CreateProperty/PropertyDetails.jsx
@@ -1,8 +1,8 @@
 import React from "react";
-import { Button, Input, RadioButton, StepperComponent } from "../../common";
 import { Form, Formik } from "formik";
 import * as yup from "yup";
-import { STEPS } from "../../../utils/constants";
+import { Button, Input, RadioButton, StepperComponent } from "../../../common";
+import { STEPS } from "../../../../utils/constants";
 
 const propertyDetailSchema = yup.object({
   subPropertyType: yup.string().required("*required"),
