+++ src/components/pages/index.jsx
@@ -1,8 +1,7 @@
 import Home from "./home/Home";
 import Login from "./login/Login";
 import Register from "./register/RegisterUser";
-import Admin from "./Admin/Admin";
 import ErrorPage from "./ErrorPage/ErrorPage";
 import RegisterBuilder from "./register/RegisterBuilder";
 
-export { Home, Login, Register, Admin, ErrorPage, RegisterBuilder };
+export { Home, Login, Register, ErrorPage, RegisterBuilder };
