+++ src/components/pages/user/SavedProperties/SavedProperties.jsx
@@ -1,9 +1,8 @@
 import React, { useEffect, useState } from "react";
 import { useSelector } from "react-redux";
-import PropertyCard from "../../common/PropertyCard";
-import { Button } from "../../common";
 import { useNavigate } from "react-router";
-import { getAllProperties } from "../../../utils/axiosGloableInstance";
+import { getAllProperties } from "../../../../utils/axiosGloableInstance";
+import { Button, PropertyCard } from "../../../common";
 
 const SavedProperties = () => {
   const navigate = useNavigate();
